+++ packages/ice-npm-utils/src/__tests__/index.ts
@@ -20,6 +20,9 @@ const defaultRegistry = 'https://registry.npm.taobao.org';
 // 也有可能返回这个源
 const cnpmResponseRegistry = 'https://registry.nlark.com' ;
 
+// 或者是这个源
+const npmMirrorRegistry = 'https://registry.npmmirror.com';
+
 jest.setTimeout(10 * 1000);
 
 
@@ -126,8 +129,8 @@ test('getNpmTarball', () => {
   return getNpmTarball('ice-npm-utils', '1.0.0').then((tarball) => {
     console.log('getNpmTarball ice-npm-utils', tarball);
     expect(
-      tarball === `${defaultRegistry}/ice-npm-utils/download/ice-npm-utils-1.0.0.tgz`
-      || tarball === `${cnpmResponseRegistry}/ice-npm-utils/download/ice-npm-utils-1.0.0.tgz`
+      [defaultRegistry, cnpmResponseRegistry, npmMirrorRegistry]
+        .some(registry => tarball === `${registry}/ice-npm-utils/download/ice-npm-utils-1.0.0.tgz`)
     ).toBeTruthy();
   });
 });
@@ -136,8 +139,8 @@ test('getNpmTarball should get latest version', () => {
   return getNpmTarball('http').then((tarball) => {
     console.log('getNpmTarball http', tarball);
     expect(
-      tarball === `${defaultRegistry}/http/download/http-0.0.1-security.tgz`
-      || tarball === `${cnpmResponseRegistry}/http/download/http-0.0.1-security.tgz`
+      [defaultRegistry, cnpmResponseRegistry, npmMirrorRegistry]
+        .some(registry => tarball === `${registry}/http/download/http-0.0.1-security.tgz`)
     ).toBeTruthy();
   });
 });