+++ packages/build-plugin-component/CHANGELOG.md
@@ -1,9 +1,18 @@
 # Changelog
 
 
-## 1.7.1
+## 1.8.2
 
 - [chore] update miniapp-compile-config version to 0.2.x
+
+## 1.8.1
+
+- [feat] support babelPlugins in `type: rax`.
+- [fix] provide consistency of `define` between rax & react ([#235](https://github.com/ice-lab/iceworks-cli/issues/235))
+
+## 1.8.0
+
+- [feat] use `disableGenerateStyle` to bypass `style.js` ([#4705](https://github.com/alibaba/ice/issues/4705))
 
 ## 1.7.0